Added story for media gallery component
